Move Response class into Audits module
This is part of a refactor to namespace our classes better.
