+++ app/decorators/audits/response_decorator.rb
@@ -1,0 +1,6 @@
+module Audits
+  class ResponseDecorator < Draper::Decorator
+    delegate_all
+    decorates_association :question
+  end
+end
