--- app/decorators/response_decorator.rb
@@ -1,4 +1,0 @@
-class ResponseDecorator < Draper::Decorator
-  delegate_all
-  decorates_association :question
-end
